Major cleanup.
Chrono and filesystem plugins are moved into the core, as they do not
depend on anything except it, so there's no point in keeping them
separate.

Perl plugin is removed, it was abandoned long ago, and now we don't need
language support plugins at all.

Added .install debhelper files for blastbeat and urlfetch so that these
plugins are actually installed. These plugins are not used in production
yet, so it wasn't critical that their packages were empty previously.