feat(node): tagged events
This commit allows to wire an event with the special slave by
specifying that slave’s id. It will be spawned if required and if there
are no limitations like exceeded pool limit. All events with some id
are guaranteed to be routed to slaves with the same id.

If a slave for some reasons reject the assignment the client will be
notified about it, and that event will be popped from the queue.